Import from agent (#2)
* add files from agent

* update requirements.txt

* update sdk version

+++ src/main.py
@@ -77,7 +77,7 @@ for dataset_name in datasets_names:
         continue
 
 sly.fs.remove_dir(dir_=g.STORAGE_DIR)
-if g.REMOVE_SOURCE:
+if g.REMOVE_SOURCE and not g.IS_ON_AGENT:
     g.api.file.remove(team_id=g.TEAM_ID, path=g.INPUT_PATH)
     source_dir_name = g.INPUT_PATH.lstrip("/").rstrip("/")
     sly.logger.info(
